fix loading state in database pages
